recompose is not longer a dependency (now a dev dependency)
* reducer file split into utils and reducers file
* mocha options moved to mocha.opts file
* tests folder renamed to test
