+++ .babelrc
@@ -9,18 +9,23 @@
     }]
   ],
   "plugins": [
-    ["lodash", { "id": ["lodash", "recompose"]}],
+    ["lodash", { "id": ["lodash"]}],
     "add-module-exports",
     "transform-object-rest-spread",
     "transform-object-assign",
-    "transform-class-properties"
+    "transform-class-properties",
+    "transform-export-extensions"
   ],
   "env": {
     "es": {
-      "comments": false
+      "comments": false,
+      "plugins": [
+      ]
     },
     "commonjs": {
-      "comments": false
+      "comments": false,
+      "plugins": [
+      ]
     },
     "test": {
       "plugins": [
