+++ package.json
@@ -9,10 +9,10 @@
   "unpkg": "dist/react-redux-firebase.min.js",
   "scripts": {
     "clean": "rimraf es lib dist coverage",
-    "lint": "eslint src/** tests/**",
+    "lint": "eslint src/** test/**",
     "lint:fix": "npm run lint -- --fix",
-    "test": "mocha -R spec ./tests/setup.js ./tests/** --recursive --require babel-register",
-    "test:cov": "istanbul cover ./node_modules/mocha/bin/_mocha -- ./tests/** --recursive --report lcov --require babel-register",
+    "test": "mocha -R spec ./test/unit/**",
+    "test:cov": "istanbul cover ./node_modules/mocha/bin/_mocha ./test/unit/**",
     "codecov": "cat coverage/lcov.info | codecov",
     "build:commonjs": "cross-env BABEL_ENV=commonjs babel src --out-dir lib",
     "build:es": "cross-env BABEL_ENV=es babel src --out-dir es",
@@ -58,8 +58,7 @@
   "dependencies": {
     "hoist-non-react-statics": "^2.3.1",
     "lodash": "^4.17.4",
-    "prop-types": "^15.5.10",
-    "recompose": "^0.26.0"
+    "prop-types": "^15.5.10"
   },
   "peerDependencies": {
     "react": "^0.14.6 || ^15.0.0-0 || ^16.0.0-0"
@@ -73,6 +72,7 @@
     "babel-plugin-lodash": "^3.2.11",
     "babel-plugin-transform-async-to-generator": "^6.24.1",
     "babel-plugin-transform-class-properties": "^6.23.0",
+    "babel-plugin-transform-export-extensions": "^6.22.0",
     "babel-plugin-transform-object-assign": "^6.22.0",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-runtime": "^6.23.0",
@@ -101,14 +101,15 @@
     "firebase": "^4.1.3",
     "firebase-server": "^0.10.1",
     "gitbook-cli": "^2.3.0",
-    "istanbul": "^1.1.0-alpha.1",
+    "istanbul": "1.1.0-alpha.1",
     "jsdom": "^9.12.0",
     "mocha": "^3.2.0",
     "react": "^15.4.2",
     "react-addons-test-utils": "^15.4.2",
     "react-dom": "^15.4.2",
+    "recompose": "^0.26.0",
     "redux": "3.6.0",
-    "redux-firestore": "^0.2.0",
+    "redux-firestore": "latest",
     "rimraf": "^2.6.2",
     "sinon": "^2.1.0",
     "sinon-chai": "^2.9.0",
