+++ src/firestoreConnect.js
@@ -1,9 +1,8 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import hoistStatics from 'hoist-non-react-statics'
-import { wrapDisplayName } from 'recompose'
-import { createCallable } from './utils'
 import { isEqual } from 'lodash'
+import hoistStatics from 'hoist-non-react-statics'
+import { createCallable, wrapDisplayName } from './utils'
 
 /**
  * @name createFirestoreConnect
