+++ src/utils/index.js
@@ -7,3 +7,20 @@ export { getEventsFromInput } from './events'
  * @param {Function|Object|Array|String} Callable function or value of return for new function
  */
 export const createCallable = f => isFunction(f) ? f : () => f
+
+export const getDisplayName = Component => {
+  if (typeof Component === 'string') {
+    return Component
+  }
+
+  if (!Component) {
+    return undefined
+  }
+
+  return Component.displayName || Component.name || 'Component'
+}
+
+export default getDisplayName
+
+export const wrapDisplayName = (BaseComponent, hocName) =>
+  `${hocName}(${getDisplayName(BaseComponent)})`
