+++ test/mocha.opts
@@ -1,0 +1,5 @@
+--require babel-core/register
+--require ./test/setup
+--recursive
+--colors
+--report lcov
