+++ test/unit/actions/auth.spec.js
