+++ test/unit/actions/query.spec.js
