+++ test/unit/actions/storage.spec.js
