+++ test/unit/enhancer.spec.js
