+++ test/unit/firebaseConnect.spec.js
