+++ test/unit/firestoreConnect.spec.js
