+++ test/unit/library.spec.js
