+++ test/unit/utils/auth.spec.js
