+++ test/unit/utils/events.spec.js
