+++ test/unit/utils/query.spec.js
