+++ test/unit/withFirebase.spec.js
